dialogue-cse/requirements.txt
jieba
tqdm
emoji
xlsxwriter
oss2
plotly
pandas==1.1.1
numpy==1.17.4
scipy
scikit-learn==0.23.2
requests==2.23.0
tensorflow-gpu==1.15.0
protobuf==3.13.0
xlrd==1.2.0
nltk==3.6.7